P4-3096 commit wiring in new controller flow into integration tests.
