+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/config/FilterConfiguration.kt
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import uk.gov.justice.digital.hmpps.pecs.jpc.controller.HtmlController
 import uk.gov.justice.digital.hmpps.pecs.jpc.controller.MaintainSupplierPricingController
-import uk.gov.justice.digital.hmpps.pecs.jpc.controller.ManagePriceCatalogueController
+import uk.gov.justice.digital.hmpps.pecs.jpc.controller.ManageJourneyPriceCatalogueController
 import uk.gov.justice.digital.hmpps.pecs.jpc.controller.ManageSchedule34LocationsController
 import uk.gov.justice.digital.hmpps.pecs.jpc.controller.MapFriendlyLocationController
 import uk.gov.justice.digital.hmpps.pecs.jpc.filter.ChooseSupplierFilter
@@ -21,7 +21,7 @@ class FilterConfiguration {
         *MaintainSupplierPricingController.routes(),
         *ManageSchedule34LocationsController.routes(),
         *MapFriendlyLocationController.routes(),
-        *ManagePriceCatalogueController.routes()
+        *ManageJourneyPriceCatalogueController.routes()
       )
     }
 }
