+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/controller/MapFriendlyLocationController.kt
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.SessionAttributes
 import org.springframework.web.servlet.mvc.support.RedirectAttributes
 import org.springframework.web.util.UriComponentsBuilder
-import uk.gov.justice.digital.hmpps.pecs.jpc.controller.HtmlController.Companion.SEARCH_JOURNEYS_RESULTS_URL
+import uk.gov.justice.digital.hmpps.pecs.jpc.controller.ManageJourneyPriceCatalogueController.Companion.SEARCH_JOURNEYS_RESULTS_URL
 import uk.gov.justice.digital.hmpps.pecs.jpc.controller.MapFriendlyLocationController.Companion.LOCATION_ORIGIN_SESSION_ATTRIBUTE
 import uk.gov.justice.digital.hmpps.pecs.jpc.controller.constraints.ValidDuplicateLocation
 import uk.gov.justice.digital.hmpps.pecs.jpc.domain.location.LocationType
