+++ src/main/resources/templates/layout/base.html
@@ -31,7 +31,7 @@
                                 <a class="moj-primary-navigation__link" href="/dashboard" aria-current="page">Move summary</a>
                             </li>
                             <li class="moj-primary-navigation__item">
-                                <a class="moj-primary-navigation__link" href="/search-journeys">Manage Journey Price Catalogue</a>
+                                <a class="moj-primary-navigation__link" href="/manage-journey-price-catalogue">Manage Journey Price Catalogue</a>
                             </li>
                             <li class="moj-primary-navigation__item">
                                 <a class="moj-primary-navigation__link" href="/search-locations">Manage Locations</a>
