+++ src/main/resources/templates/manage-journey-price-catalogue.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+
+<html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+      layout:decorate="~{layout/base}">
+
+<head>
+  <title>Calculate Journey Variable Payments - GOV.UK</title>
+</head>
+
+<body>
+
+<div class="govuk-width-container" layout:fragment="content">
+  <div class="mv4">
+    <h1 class="govuk-heading-xl mv2" th:inline="text">
+      Manage Journey Price Catalogue [[${contractualYearStart}]] to [[${contractualYearEnd}]]
+    </h1>
+  </div>
+
+  <div class="govuk-width-container">
+    <a href="/search-journeys" class="govuk-heading-s govuk-link">Find journeys</a>
+    <span>Search and edit individual journey prices</span>
+  </div>
+</div>
+
+</body>
+
+</html>
