+++ src/main/resources/templates/search-journeys.html
@@ -47,7 +47,7 @@
       <button id="find-journeys" class="govuk-button mv3 mr3" data-module="govuk-button">
         Find Journeys
       </button>
-      <a href="/dashboard" class="govuk-link">Cancel</a>
+      <a href="/manage-journey-price-catalogue" class="govuk-link">Cancel</a>
     </div>
   </form>
